+++ CHANGELOG.rst
@@ -95,6 +95,7 @@ People and network changes
 
 Other changes
 ~~~~~~~~~~~~~
+- Boolean arrays have new methods ``true()``, ``false()``, and ``split()``, which return the UIDs for the ``True`` values (alias to ``arr.uids``), ``False`` values, and both sets of values, respectively. ``ss.bernoulli.split()`` has been added as an alias of ``ss.bernoulli.filter(both=True)``.
 - All inputs to a sim are now copied by default. To disable, use ``ss.Sim(..., copy_inputs=False)``.
 - There is a new ``Plugin`` class, which contains shared logic for Interventions and Analyzers. It has a ``from_func()``, which will generate an intervention/analyzer from a function.
 - Diseases no longer have a default value of ``beta=1`` assigned; beta must be defined explicitly if being used.